add auto_color option
try to sort out when we decide an attribute should
try to pick a color automatically or not.

auto_color=False (the default) means dont try to get one
for attributes unless specified in some fashion
(in color_groups or default_color_attr).

Reduce the defaults to not doing much, so we dont have
to override them.