Add ex if vaguely named thread for 'threaded'
This is to demonstrate how the 'by thread' (tid) log colorization
can be useful when the thread name is vague. For example, it
is not uncommon for an app to end up starting multiple threads
that all have the same thread name, often something like 'helper'.

For cases like that, if you want all of the 'helper' threads to
use the same color, you can use 'threadName' for the default_color_attr.

If you want to see each 'helper' thread instance in a different color,
you can use 'thread' (tid) attribute for default_color_attr.