Make sure Mappers with default args use default attr
If a default_color_by_attr wasn't getting passed in
it's default is None, but self.default_color_by_attr
will use a default in that case.

But the BaseColorMapper init was checking if the
init local default_color_by_args was truthy instead,
so wasn't setting the default color group. So
a TermFormatter() would end up not applying any colors
because there was nothing added to color_groups